Fix calculation of maximum values for swallow parts.
Due my last change, the code was broken by edjes without a group
min/max, this happens because edje_object_size_max_get() returns
100000 for these objects, and this was being used as object maximum
size.

Current fix is a hack: just check for this value, now known as
EDJE_INF_MAX_*, but the real solution would be to return 0 (or -1) and
check for it in other parts of the code, but it's harder to get right.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e17/libs/edje@32123 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33