Avoid using anonymous for often used structs.
My last patch fixed the compile problem, but really, we should use the
same struct for things that look/work the same, like rectangle,
position, etc.

This new patch brings that change back and add some named structs
(also typedef'ed so it conforms to E naming schema).

This commit just changes the header and adds an example of
benefit. Later I'll provide a more intrusive patch that reorganize
structures to make even better use of this.