Interpolate text font size.
Often requested for animations that want to grow or shrink text
quickly. This is a faster alternative to using geometry with rel1/rel2
and "text.fit: 1 1;" since it does not need to figure out which size
fits better into that object.

I tested and it does not seem to introduce any regression. Also
checked with scale and text_class variations.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/edje@38958 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33