let's stop smoking crack while coding :-P
dudes, WTF is that look to check for numbers, then use atoi()??? Use
strtol() and let it do both in one pass. Also, just do these for part
types that make sense. The loop using child could result in unused
variables in lots of cases, returning the wrong value when not found.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/edje@50757 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33