edje: automatically emit ['edje,change,file', 'edje'] when the file change on disk.
For the moment only edje_player use it. This means that when used with
edje_watch, you don't need any more to type any kind of command line
when you are testing value in your theme. As a side effect, this means
that their is a real use case to make edje_cc faster !



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/edje@70890 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33