From: Bluezery <ohpowel@gmail.com> Subject: [E-devel] [Patch][edje] Prevent duplicated inherition of "after"
When inherting group, "after" of "program" section is also inherited.
But duplicated name of "after" can be also inherited.
I think that multiple "after" can be used but duplicated should be
not.  because this seems to be useless.
Is there case that duplicated after usage is used?
Duplicated "target" is prevented.  I just copy those codes.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/edje@80565 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33